--- nft-v2/test/03-approve.test.ts
@@ -1,18 +1,0 @@
-// import {
-// aliceActor,
-// bobActor,
-// canisterOwnerActor,
-// johnActor,
-// } from "./setup";
-import test from "ava";
-
-// const normalActors = [aliceActor, bobActor, johnActor];
-// const allActors = [...normalActors, canisterOwnerActor];
-
-test("OK", async t => {
-  t.true(true);
-});
-
-test("ERROR", async t => {
-  t.true(true);
-});
